+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Exmaralda exporter would export the same file several times depending on the
+  token number.
+
 ## [0.39.0] - 2025-08-25
 
 ### Added